add share button at token page
